+++ packages/components/src/components/link-tile/link-tile-utils.ts
@@ -1,5 +1,5 @@
-import { TILE_WEIGHTS, AllowedTypes, TILE_SIZES, TILE_ASPECT_RATIOS, TILE_ALIGNS } from '../../utils';
-import type { TileAlign, TileAspectRatio, TileSize } from '../../utils';
+import { TILE_WEIGHTS, AllowedTypes, TILE_SIZES, TILE_ASPECT_RATIOS, TILE_ALIGNS, THEMES } from '../../utils';
+import type { TileAlign, TileAspectRatio, TileSize, TileBackground } from '../../utils';
 import type { LinkTarget } from '../../utils/link-button/link-target';
 import type { LinkAriaAttribute } from '../link/link-utils';
 import type { PropTypes } from '../../utils/validation/validateProps';
@@ -13,6 +13,7 @@ export type LinkTileTarget = LinkTarget;
 export type LinkTileAriaAttribute = LinkAriaAttribute;
 export type LinkTileAspectRatio = TileAspectRatio;
 export type LinkTileSize = TileSize;
+export type LinkTileBackground = TileBackground;
 export type LinkTileAlign = TileAlign;
 export const LINK_TILE_WEIGHTS = [...TILE_WEIGHTS, 'semibold'] as const;
 export type LinkTileWeight = (typeof LINK_TILE_WEIGHTS)[number];
@@ -25,6 +26,7 @@ type CommonButtonAndLinkTileProps = {
 
 export const sharedTilePropTypes: Omit<CommonButtonAndLinkTileProps, 'aria' | 'weight'> = {
   size: AllowedTypes.breakpoint<TileSize>(TILE_SIZES),
+  background: AllowedTypes.oneOf<TileBackground>(THEMES),
   aspectRatio: AllowedTypes.breakpoint<TileAspectRatio>(TILE_ASPECT_RATIOS),
   label: AllowedTypes.string,
   description: AllowedTypes.string,